// types-env.d.ts
type TokenData = {
  expirationDate: string;
  token: string;
  userId: number;
  uuid: string;
  idIndividu: number;
  etablissementId: number;
  userName: string;
};

type StudentInfo = {
  anneeBac: string;
  dateNaissance: string;
  id: number;
  libelleSerieBac: string;
  matricule: string;
  moyenneBac: string;
  nin: string;
  nomAr: string;
  nomFr: string;
  prenomAr: string;
  prenomFr: string;
  refCodeSerieBac: string;
  refCodeWilayaBac: string;
  uuid: string;
};